Move request handling logic to the logic package and rename example classes
