--- src/main/java/pl/put/poznan/transformer/logic/TextTransformer.java
@@ -1,18 +1,0 @@
-package pl.put.poznan.transformer.logic;
-
-/**
- * This is just an example to show that the logic should be outside the REST service.
- */
-public class TextTransformer {
-
-    private final String[] transforms;
-
-    public TextTransformer(String[] transforms){
-        this.transforms = transforms;
-    }
-
-    public String transform(String text){
-        // of course, normally it would do something based on the transforms
-        return text.toUpperCase();
-    }
-}
